+++ environment/identity.ts
@@ -1,7 +1,7 @@
 import { ResourceGroup } from '@pulumi/azure-native/resources'
 import { Application, ServicePrincipal, ServicePrincipalPassword } from '@pulumi/azuread'
 
-import * as environment from '~/helpers/environment'
+import * as environment from '~/shared/environment'
 
 export = async () => {
   const application = new Application( `${environment.name}-environment-app`, { displayName: `${environment.name}-environment-app` } )
