+++ shared/azure.ts
